refactor(winglet): resolve zero-peer findings with organ moves and scoped grants (#330)
- withErrorBoundary: 내부 컴포넌트(ErrorBoundary·FallbackMessage)를 components/ organ으로
  이동하고 계약 문서(INTENT/DETAIL)를 선행 작성
- 구조 검증 설정: entry가 공개 export하는 flat 유닛만 남는 fractal 11곳에
  zero-peer scoped exempt 부여 (#333의 flat-canonical 결정을 fractal 단위로 확장)
- function entry의 rateLimit 딥 import 2건을 자식 entry point 경유로 수정
  (import-boundary 해소)

검증: 구조 검증 엔진 기준 winglet zero-peer 28→0, import-boundary 2→0;
@winglet/* typecheck·build·테스트 통과
