+++ packages/winglet/common-utils/src/utils/function/index.ts
@@ -3,5 +3,9 @@ export {
   type TrackableHandlerFunction,
   type TrackableHandlerOptions,
 } from './enhance/getTrackableHandler';
-export { debounce, type DebouncedFn } from './rateLimit/debounce';
-export { throttle, type ThrottledFn } from './rateLimit/throttle';
+export {
+  debounce,
+  type DebouncedFn,
+  throttle,
+  type ThrottledFn,
+} from './rateLimit';
